--- src/bets/dto/create-answer-res-dto.ts
@@ -1,9 +1,0 @@
-import { ApiProperty } from '@nestjs/swagger';
-
-export class CreateAnswerResDto {
-  @ApiProperty({
-    description: 'answer id',
-    type: Number,
-  })
-  answerId!: number;
-}
