+++ src/bets/dto/get-participants.dto.ts
