+++ src/bets/dto/get-total-prediction.dto.ts
